Upload the documentation files
The "presentation.pdf" are slides containing an overview of the project. I used it to present it to my tutors and to a competitions organized by the company (CFZ Cobots).

The "report.pdf" is a report containing deeper explanations of which things I did and why I choosed them instead of some others alternatives I also mention.